pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "huaqi"
version = "0.1.0"
description = "个人 AI 同伴系统 - Personal AI Companion"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "连子蒙", email = "your-email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "prompt-toolkit>=3.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    # LangGraph Agent 架构
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "aiosqlite>=0.17.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    # 定时任务
    "apscheduler>=4.0.0a",
    # 向量存储
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.0",
    # 数据获取
    "aiohttp>=3.9.0",
    "feedparser>=6.0.0",
    # 配置热重载
    "watchdog>=4.0.0",
    # 结构化日志
    "structlog>=24.0.0",
    # CLI 语法高亮
    "pygments>=2.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
huaqi = "cli:app"

[tool.setuptools]
py-modules = ["cli"]

[tool.setuptools.packages.find]
where = ["."]
include = ["huaqi*", "cli"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
